添加删除二次确认

+++ View/FavouritePlayListPage.xaml.cs
@@ -111,10 +111,13 @@ namespace WinUIMusicPlayer.View
             ViewModel.PlayMenuItem_Click(uniqueSelectedMusics);
         }
 
-        private void DeleteMenuItem_Click(object sender, RoutedEventArgs e)
+        private async void DeleteMenuItem_Click(object sender, RoutedEventArgs e)
         {
-            IEnumerable<Music> uniqueSelectedMusics = GetUniqueSelectedItems();
-            ViewModel.DeleteMenuItem_Click(uniqueSelectedMusics);
+            if (await ViewModel.IsDeleteFromDisk())
+            {
+                IEnumerable<Music> uniqueSelectedMusics = GetUniqueSelectedItems();
+                ViewModel.DeleteMenuItem_Click(uniqueSelectedMusics);
+            }            
         }
 
         private void SetAsFavoriteMenuItem_Click(object sender, RoutedEventArgs e)
