+++ View/SongListPage.xaml.cs
@@ -88,10 +88,13 @@ namespace WinUIMusicPlayer.View
             await ViewModel.ConvertAudio_Click(uniqueSelectedMusics, menuItem);
         }
 
-        private void DeleteMenuItem_Click(object sender, RoutedEventArgs e)
+        private async void DeleteMenuItem_Click(object sender, RoutedEventArgs e)
         {
-            IEnumerable<Music> uniqueSelectedMusics = GetUniqueSelectedItems();
-            ViewModel?.DeleteMenuItem_Click(uniqueSelectedMusics);
+            
+            if (await ViewModel.IsDeleteFromDisk()) {
+                IEnumerable<Music> uniqueSelectedMusics = GetUniqueSelectedItems();
+                ViewModel?.DeleteMenuItem_Click(uniqueSelectedMusics);
+            }           
         }
 
         private async void SetAsFavoriteMenuItem_Click(object sender, RoutedEventArgs e)
