+++ ViewModel/SongCollectionViewModel.cs
@@ -221,6 +221,14 @@ namespace WinUIMusicPlayer.ViewModel
             }
         }
 
+        public async Task<bool> IsDeleteFromDisk() {
+            if (_parentPage == null)
+            {
+                return false;
+            }
+            return await _parentPage.AreUSureDeleteFromDisk();
+        }
+
         private void GetCover() {
             _ = Task.Run(async () =>
             {
@@ -369,10 +377,11 @@ namespace WinUIMusicPlayer.ViewModel
                 for (int i = MusicList.Count - 1; i >= 0; i--)
                 {
                     if (uniqueSelectedMusics.Contains(MusicList[i]))
-                    {
-                        await MusicDatabaseService.RemoveMusic((MusicList[i].Id));
-                        ToolUtils.DeleteFileFromDisk((MusicList[i].Path));
-                        MusicList.RemoveAt(i);
+                    {                        
+                        if (ToolUtils.DeleteFileFromDisk(MusicList[i].Path)) {
+                            await MusicDatabaseService.RemoveMusic(MusicList[i].Id);
+                            MusicList.RemoveAt(i);
+                        }                       
                     }
                 }
                 //foreach (Music item in uniqueSelectedMusics)
@@ -385,10 +394,11 @@ namespace WinUIMusicPlayer.ViewModel
             else
             {
                 if (SelectedMusic != null)
-                {
-                    await MusicDatabaseService.RemoveMusic(SelectedMusic.Id);
-                    ToolUtils.DeleteFileFromDisk(SelectedMusic.Path);
-                    MusicList.Remove(SelectedMusic);
+                {                    
+                    if (ToolUtils.DeleteFileFromDisk(SelectedMusic.Path)) {
+                        await MusicDatabaseService.RemoveMusic(SelectedMusic.Id);
+                        MusicList.Remove(SelectedMusic);
+                    }
                 }
             }
